a2ui-demo/samples/agent/chainlit/pyproject.toml
[project]
name = "chainlit-gemini-chat"
version = "0.1.0"
description = "Chainlit + Gemini chat application - for comparison with A2UI Simple Chat"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.0.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
